Added support for better handling of transitive dependencies

+++ src/NuGetForUnity/Editor/NugetHelper.cs
@@ -67,6 +67,8 @@ namespace NugetForUnity
         /// </summary>
         private static readonly List<NugetPackageSource> packageSources = new List<NugetPackageSource>();
 
+        public static event Action OnInstalledPackagesChanged;
+
         /// <summary>
         ///     The dictionary of currently installed <see cref="NugetPackage" />s keyed off of their ID string.
         /// </summary>
@@ -518,7 +520,7 @@ namespace NugetForUnity
             }
         }
 
-        private static bool IsAlreadyImportedInEngine(NugetPackageIdentifier package, bool log = true)
+        internal static bool IsAlreadyImportedInEngine(NugetPackageIdentifier package, bool log = true)
         {
             var alreadyImportedLibs = UnityPreImportedLibraryResolver.GetAlreadyImportedLibs();
             var isAlreadyImported = alreadyImportedLibs.Contains(package.Id);
@@ -722,33 +724,62 @@ namespace NugetForUnity
         {
             foreach (var package in packagesToUninstall)
             {
-                Uninstall(package);
+                Uninstall(package, false);
             }
+
+            AssetDatabase.Refresh();
         }
 
         /// <summary>
         ///     "Uninstalls" the given package by simply deleting its folder.
         /// </summary>
         /// <param name="package">The NugetPackage to uninstall.</param>
-        /// <param name="refreshAssets">True to force Unity to refesh its Assets folder.  False to temporarily ignore the change.  Defaults to true.</param>
+        /// <param name="refreshAssets">True to force Unity to refresh its Assets folder.  False to temporarily ignore the change.  Defaults to true.</param>
         public static void Uninstall(NugetPackageIdentifier package, bool refreshAssets = true)
         {
             LogVerbose("Uninstalling: {0} {1}", package.Id, package.Version);
 
+            var foundPackage = package as NugetPackage ?? GetSpecificPackage(package);
+
             // update the package.config file
-            PackagesConfigFile.RemovePackage(package);
+            if (!PackagesConfigFile.RemovePackage(foundPackage))
+            {
+                return;
+            }
+
             PackagesConfigFile.Save(PackagesConfigFilePath);
 
-            var packageInstallDirectory = Path.Combine(NugetConfigFile.RepositoryPath, $"{package.Id}.{package.Version}");
+            var packageInstallDirectory = Path.Combine(NugetConfigFile.RepositoryPath, $"{foundPackage.Id}.{foundPackage.Version}");
             DeleteDirectory(packageInstallDirectory);
 
-            var metaFile = Path.Combine(NugetConfigFile.RepositoryPath, $"{package.Id}.{package.Version}.meta");
+            var metaFile = Path.Combine(NugetConfigFile.RepositoryPath, $"{foundPackage.Id}.{foundPackage.Version}.meta");
             DeleteFile(metaFile);
 
-            var toolsInstallDirectory = Path.Combine(AbsoluteProjectPath, "Packages", $"{package.Id}.{package.Version}");
+            var toolsInstallDirectory = Path.Combine(AbsoluteProjectPath, "Packages", $"{foundPackage.Id}.{foundPackage.Version}");
             DeleteDirectory(toolsInstallDirectory);
 
-            installedPackages?.Remove(package.Id);
+            installedPackages?.Remove(foundPackage.Id);
+
+            if (installedPackages != null && installedPackages.Count > 0)
+            {
+                var frameworkGroup = GetBestDependencyFrameworkGroupForCurrentSettings(foundPackage);
+                foreach (var dependency in frameworkGroup.Dependencies)
+                {
+                    var actualData = PackagesConfigFile.Packages.Find(pkg => pkg.Id == dependency.Id);
+                    if (actualData == null || actualData.IsManuallyInstalled)
+                    {
+                        continue;
+                    }
+
+                    var hasMoreParents = installedPackages.Values.Select(GetBestDependencyFrameworkGroupForCurrentSettings)
+                        .Any(frameworkGrp => frameworkGrp.Dependencies.Any(dep => dep.Id == dependency.Id));
+
+                    if (!hasMoreParents)
+                    {
+                        Uninstall(dependency, false);
+                    }
+                }
+            }
 
             if (refreshAssets)
             {
@@ -766,6 +797,7 @@ namespace NugetForUnity
         {
             LogVerbose("Updating {0} {1} to {2}", currentVersion.Id, currentVersion.Version, newVersion.Version);
             Uninstall(currentVersion, false);
+            newVersion.IsManuallyInstalled = currentVersion.IsManuallyInstalled;
             return InstallIdentifier(newVersion, refreshAssets);
         }
 
@@ -804,6 +836,12 @@ namespace NugetForUnity
             EditorUtility.ClearProgressBar();
         }
 
+        public static void UpdatePackageConfig(NugetPackage package)
+        {
+            PackagesConfigFile.AddPackage(package);
+            PackagesConfigFile.Save(PackagesConfigFilePath);
+        }
+
         /// <summary>
         ///     Updates the dictionary of packages that are actually installed in the project based on the files that are currently installed.
         /// </summary>
@@ -826,19 +864,29 @@ namespace NugetForUnity
             // loops through the packages that are actually installed in the project
             if (Directory.Exists(NugetConfigFile.RepositoryPath))
             {
-                // a package that was installed via NuGet will have the .nupkg it came from inside the folder
-                var nupkgFiles = Directory.GetFiles(NugetConfigFile.RepositoryPath, "*.nupkg", SearchOption.AllDirectories);
-                foreach (var nupkgFile in nupkgFiles)
+                void AddPackageToInstalled(NugetPackage package)
                 {
-                    var package = NugetPackage.FromNupkgFile(nupkgFile);
                     if (!installedPackages.ContainsKey(package.Id))
                     {
+                        var actualData = PackagesConfigFile.Packages.Find(pkg => pkg.Id == package.Id);
+                        if (actualData != null)
+                        {
+                            package.IsManuallyInstalled = actualData.IsManuallyInstalled;
+                        }
                         installedPackages.Add(package.Id, package);
                     }
                     else
                     {
                         Debug.LogErrorFormat("Package is already in installed list: {0}", package.Id);
                     }
+                }
+
+                // a package that was installed via NuGet will have the .nupkg it came from inside the folder
+                var nupkgFiles = Directory.GetFiles(NugetConfigFile.RepositoryPath, "*.nupkg", SearchOption.AllDirectories);
+                foreach (var nupkgFile in nupkgFiles)
+                {
+                    var package = NugetPackage.FromNupkgFile(nupkgFile);
+                    AddPackageToInstalled(package);
                 }
 
                 // if the source code & assets for a package are pulled directly into the project (ex: via a symlink/junction) it should have a .nuspec defining the package
@@ -846,19 +894,14 @@ namespace NugetForUnity
                 foreach (var nuspecFile in nuspecFiles)
                 {
                     var package = NugetPackage.FromNuspec(NuspecFile.Load(nuspecFile));
-                    if (!installedPackages.ContainsKey(package.Id))
-                    {
-                        installedPackages.Add(package.Id, package);
-                    }
-                    else
-                    {
-                        Debug.LogErrorFormat("Package is already in installed list: {0}", package.Id);
-                    }
+                    AddPackageToInstalled(package);
                 }
             }
 
             stopwatch.Stop();
             LogVerbose("Getting installed packages took {0} ms", stopwatch.ElapsedMilliseconds);
+
+            OnInstalledPackagesChanged?.Invoke();
         }
 
         /// <summary>
@@ -1087,6 +1130,7 @@ namespace NugetForUnity
 
             if (foundPackage != null)
             {
+                foundPackage.IsManuallyInstalled = package.IsManuallyInstalled;
                 return Install(foundPackage, refreshAssets);
             }
 
@@ -1178,7 +1222,7 @@ namespace NugetForUnity
                             "Can't find a matching dependency group for the NuGet Package {0} {1} that has a TargetFramework supported by the current Unity Scripting Backend. The NuGet Package supports the following TargetFramework's: {2}",
                             package.Id,
                             package.Version,
-                            string.Join(", ", package.Dependencies.Select(dependeny => dependeny.TargetFramework)));
+                            string.Join(", ", package.Dependencies.Select(dependency => dependency.TargetFramework)));
                     }
                     else if (frameworkGroup != null)
                     {
@@ -1251,12 +1295,11 @@ namespace NugetForUnity
                         foreach (var entry in zip.Entries)
                         {
                             var filePath = Path.Combine(baseDirectory, entry.FullName);
-                            var directory = Path.GetDirectoryName(filePath);
-                            Directory.CreateDirectory(directory);
-                            if (Directory.Exists(filePath))
-                            {
-                                continue;
-                            }
+
+                            var isDir = filePath.EndsWith("/") || filePath.EndsWith("\\");
+                            var dirName = isDir ? filePath : Path.GetDirectoryName(filePath);
+                            if (dirName != null && !Directory.Exists(dirName)) Directory.CreateDirectory(dirName);
+                            if (isDir) continue;
 
                             entry.ExtractToFile(filePath, true);
 
@@ -1269,7 +1312,7 @@ namespace NugetForUnity
                     }
 
                     // copy the .nupkg inside the Unity project
-                    File.Copy(cachedPackagePath, Path.Combine(baseDirectory, string.Format("{0}.{1}.nupkg", package.Id, package.Version)), true);
+                    File.Copy(cachedPackagePath, Path.Combine(baseDirectory, $"{package.Id}.{package.Version}.nupkg"), true);
                 }
                 else
                 {
